+++ PRISTINE/toy_dataset/data/raw_genomes/Yersinia_pseudotuberculosis/non-targets/Yersinia_pestis_1.fasta
@@ -1,4 +1,4 @@
->Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_accn|CP177922   Yersinia pestis strain 50008 chromosome, complete genome.   [Yersinia pestis 50008 | 632.1142]
+>Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_accn|CP177922   Yersinia pestis strain 50008 chromosome, complete genome.   [Yersinia pestis 50008 | 632.1142]
 ggcggactccactcgaacaagatcaataaaagaaagaaatcaggatattttgtaactcat
 cgttttgacacacgccatatggcctgtagcaagccaaggtgacatgcgaactgtttttta
 cagttgttttttactgctgattttttacgactcacattcccccacgatcccagagtagag
