+++ PRISTINE/toy_dataset/data/raw_genomes/Yersinia_pseudotuberculosis/non-targets/Yersinia_pestis_2.fasta
@@ -1,4 +1,4 @@
->Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_accn|CP177934   Yersinia pestis strain 96005 chromosome, complete genome.   [Yersinia pestis 96005 | 632.1143]
+>Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_Yersinia_pestis_accn|CP177934   Yersinia pestis strain 96005 chromosome, complete genome.   [Yersinia pestis 96005 | 632.1143]
 ggcggactccactcgaacaagatcaataaaagaaagaaatcaggatattttgtaactcat
 cgttttgacacacgccatatggcctgtagcaagccaaggtgacatgcgaactgtttttta
 cagttgttttttactgctgattttttacgactcacattcccccacgatcccagagtagag
